requirements.txt
snowflake-connector-python[pandas]==2.6.2
pytrends==4.7.3
